validator utils
