+++ backend/src/middlewares/validator.middleware.js
@@ -1,0 +1,7 @@
+import { validationResult } from "express-validator";
+
+export const middlewareValidationResult = (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) return res.status(400).json({ errors: errors.array() });
+    next();
+};
