+++ backend/src/routes/auth.router.js
@@ -1,7 +1,6 @@
 import express from "express";
-
 import { login, register } from "../controllers/auth.controller.js";
-import { loginValidator, registerValidator } from "../middlewares/validator.manager.js";
+import { loginValidator, registerValidator } from "../utils/validator.manager.js";
 
 const router = express.Router();
