+++ backend/src/routes/profile.router.js
@@ -1,5 +1,4 @@
 import express from "express";
-
 import { getProfile } from "../controllers/profile.controller.js";
 import { requireJwt } from "../middlewares/require.token.js";
